ALCS-881: Prevent User From Deleting a Decision Draft, Reconsideration, or Modification in the Middle of Decision Chain

+++ services/apps/alcs/src/alcs/application-decision/application-decision-v2/application-decision/application-decision-v2.service.ts
@@ -507,6 +507,40 @@ export class ApplicationDecisionV2Service {
       throw new ServiceNotFoundException(`Failed to find decision with uuid ${uuid}`);
     }
 
+    const decisionsWithModifiedBy = await this.appDecisionRepository.find({
+      where: {
+        applicationUuid: applicationDecision.application.uuid,
+      },
+      relations: {
+        modifiedBy: {
+          resultingDecision: true,
+          reviewOutcome: true,
+        },
+      },
+    });
+
+    const decisionsWithReconsideredBy = await this.appDecisionRepository.find({
+      where: {
+        applicationUuid: applicationDecision.application.uuid,
+      },
+      relations: {
+        reconsideredBy: {
+          resultingDecision: true,
+          reviewOutcome: true,
+        },
+      },
+    });
+
+    const reconsideredBy =
+      decisionsWithReconsideredBy.find((r) => r.uuid === applicationDecision.uuid)?.reconsideredBy || [];
+    const modifiedBy = decisionsWithModifiedBy.find((r) => r.uuid === applicationDecision.uuid)?.modifiedBy || [];
+
+    if (reconsideredBy.length > 0 || modifiedBy.length > 0) {
+      throw new ServiceValidationException(
+        `Cannot delete decision ${applicationDecision.uuid} with linked reconsiderations or modifications`,
+      );
+    }
+
     await this.decisionConditionService.remove(applicationDecision.conditions);
     applicationDecision.conditions = [];
